+++ lib/src/rendering/render_dynamic_timeline.dart
@@ -5,19 +5,16 @@ import 'package:flutter/rendering.dart';
 
 class DynamicTimelineParentData extends ContainerBoxParentData<RenderBox> {
   DynamicTimelineParentData({
-    required this.secondExtent,
+    required this.microsecondExtent,
     required this.minItemDuration,
     required this.axis,
     required this.resizable,
   });
 
-  bool get isEvent =>
-      startDateTime != null && endDateTime != null && position != null;
-
   DateTime? startDateTime;
   DateTime? endDateTime;
   int? position;
-  final double secondExtent;
+  final double microsecondExtent;
   final Duration minItemDuration;
   final Axis axis;
   final bool resizable;
@@ -239,7 +236,7 @@ class RenderDynamicTimeline extends RenderBox
   void setupParentData(covariant RenderObject child) {
     if (child.parentData is! DynamicTimelineParentData) {
       child.parentData = DynamicTimelineParentData(
-        secondExtent: intervalExtent / intervalDuration.inSeconds,
+        microsecondExtent: intervalExtent / intervalDuration.inMicroseconds,
         minItemDuration: minItemDuration,
         axis: axis,
         resizable: resizable,
@@ -271,15 +268,9 @@ class RenderDynamicTimeline extends RenderBox
       late final DateTime endDateTime;
       late final int position;
 
-      if (childParentData.isEvent) {
-        startDateTime = childParentData.startDateTime!;
-        endDateTime = childParentData.endDateTime!;
-        position = childParentData.position!;
-      } else {
-        startDateTime = firstDateTime;
-        endDateTime = firstDateTime.add(intervalDuration);
-        position = 0;
-      }
+      startDateTime = childParentData.startDateTime!;
+      endDateTime = childParentData.endDateTime!;
+      position = childParentData.position!;
 
       final childDuration = endDateTime.difference(startDateTime);