+++ lib/src/widgets/widgets.dart
@@ -1,3 +1,2 @@
 export 'dynamic_timeline.dart';
-export 'raw_timeline_item.dart';
 export 'timeline_item.dart';