+++ test/helpers/pump_app.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_test/flutter_test.dart';
+
+extension PumpApp on WidgetTester {
+  Future<void> pumpApp(Widget widget, {ThemeData? theme}) {
+    return pumpWidget(
+      MaterialApp(
+        theme: theme,
+        home: Scaffold(
+          body: widget,
+        ),
+      ),
+    );
+  }
+}