+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 0.1.1
+* fix: fix item not resizing properly
+* refactor: 
+  * delete raw_timeline_item
+  * use items instead of children
+* test: 100% test coverage
+* ci: add workflows
+  
+
 # 0.1.0+1
 
 * Readme updated.